Monty-Tuning Robot
Monty-Tuning  is the "tuned/maked" version of the original Monty robot I made in 2002.

Monty-Tuning robot
It is a robot that moves by means of a wheels differential traction and has many sensors and actuators:

Two motors for traction,
A motor for the clamp of the arm,
Ultrasounds sensor for the detection of movement,
Two sensors of light in the head for the location of light sources,
Sensors of infrared reflection in the base for pursuit of black lines,
Microphone,
Loudspeaker and
Bumpers for the detection of obstacles.

The set is governed by a microcontroller PIC-16F84 and another small microcontroller (PIC-12C50A) preprogrammed for the management of the movements of the clamp.